+++ arbiter-ios/ArbiterIOS/ArbiterIOSApp.swift
@@ -1,0 +1,12 @@
+import SwiftUI
+
+@main
+struct ArbiterIOSApp: App {
+    @State private var model = ChatViewModel()
+
+    var body: some Scene {
+        WindowGroup {
+            ChatView(model: model)
+        }
+    }
+}